renaming all task config files
