+++ human_experiments/lab_software/tomcat-baseline-tasks/tasks/affective_task/client_affective_task.py
@@ -4,7 +4,7 @@ from common import (cursor_visibility, render_blank_screen,
 from config import CLIENT_WINDOW_HEIGHT, CLIENT_WINDOW_WIDTH
 from network import receive, send
 
-from .config_affective_task import (BLANK_SCREEN_MILLISECONDS,
+from .config import (BLANK_SCREEN_MILLISECONDS,
                                     CROSS_SCREEN_MILLISECONDS,
                                     DISPLAY_AFFEC_DISCUSSION_MILLISECONDS)
