+++ human_experiments/lab_software/tomcat-baseline-tasks/tasks/ping_pong_task/client_ping_pong_task.py
@@ -7,7 +7,7 @@ from config import (BLANK_SCREEN_COUNT_DOWN_MILLISECONDS, CLIENT_WINDOW_HEIGHT,
                     CLIENT_WINDOW_WIDTH, UPDATE_RATE)
 from network import receive, send
 
-from .config_ping_pong_task import (COUNT_DOWN_MESSAGE,
+from .config import (COUNT_DOWN_MESSAGE,
                                     SHOW_SCORE_COUNT_DOWN_MILLISECONDS)
 from .utils import BALL_SIZE, WINDOW_HEIGHT, WINDOW_WIDTH, Ball, Paddle
