+++ human_experiments/lab_software/tomcat-baseline-tasks/tasks/rest_state/client_rest_state.py
@@ -3,7 +3,7 @@ from common import (cursor_visibility, render_blank_screen,
 from config import CLIENT_WINDOW_HEIGHT, CLIENT_WINDOW_WIDTH
 from network import receive, send
 
-from .config_rest_state import (BLANK_SCREEN_MILLISECONDS,
+from .config import (BLANK_SCREEN_MILLISECONDS,
                                     CROSS_SCREEN_MILLISECONDS)
 from .utils import timer
